% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastVoteR-package.R
\docType{package}
\name{fastVoteR-package}
\alias{fastVoteR}
\alias{fastVoteR-package}
\title{fastVoteR: Efficient Voting Methods for Committee Selection}
\description{
A fast 'Rcpp'-based implementation of polynomially-computable voting theory methods for committee ranking and scoring. The package includes methods such as Approval Voting (AV), Satisfaction Approval Voting (SAV), sequential Proportional Approval Voting (PAV), and sequential Phragmen's Rule. Weighted variants of these methods are also provided, allowing for differential voter influence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bblodfon.github.io/fastVoteR/}
}

}
\author{
\strong{Maintainer}: John Zobolas \email{bblodfon@gmail.com} (\href{https://orcid.org/0000-0002-3609-8674}{ORCID})

Other contributors:
\itemize{
  \item Anne-Marie George (\href{https://orcid.org/0000-0001-9232-8211}{ORCID}) [contributor]
}

}
